+++ packages/terra-clinical-result/src/terra-dev-site/doc/example/flowsheet-result-cell/FlowsheetResultCellNumericOverflow.jsx
@@ -1,8 +1,12 @@
 import React from 'react';
 import { FlowsheetResultCell } from 'terra-clinical-result/lib/index';
-import classNames from 'classnames';
 import classNamesBind from 'classnames/bind';
-import ThemeContext from 'terra-theme-context';
+import Table, {
+  Header,
+  Row,
+  Body,
+  HeaderCell,
+} from 'terra-html-table';
 import styles from '../Examples.module.scss';
 
 const cx = classNamesBind.bind(styles);
@@ -40,41 +44,24 @@ const multipleResultValues = [
   },
 ];
 
-export default () => {
-  const theme = React.useContext(ThemeContext);
-  const mockFlowsheetClassnames = classNames(
-    cx(
-      'mock-flowsheet-example',
-      theme.className,
-    ),
-  );
-
-  return (
-    <div className={mockFlowsheetClassnames}>
-      <div className={cx('mock-flowsheet-resultcolumn')}>
-        <div className={cx('mock-flowsheet-resultcolumn-cell')}>
-          <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
-        </div>
-        <div className={cx('mock-flowsheet-resultcolumn-cell')}>
-          <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
-        </div>
-      </div>
-      <div className={cx(['mock-flowsheet-resultcolumn', 'half-column'])}>
-        <div className={cx('mock-flowsheet-resultcolumn-cell')}>
-          <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
-        </div>
-        <div className={cx('mock-flowsheet-resultcolumn-cell')}>
-          <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
-        </div>
-      </div>
-      <div className={cx(['mock-flowsheet-resultcolumn', 'triple-column'])}>
-        <div className={cx('mock-flowsheet-resultcolumn-cell')}>
-          <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
-        </div>
-        <div className={cx('mock-flowsheet-resultcolumn-cell')}>
-          <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
-        </div>
-      </div>
-    </div>
-  );
-};
+export default () => (
+  <Table className={cx('mock-flowsheet-table')}>
+    <Header>
+      <HeaderCell className={cx('mock-flowsheet-resultcolumn')} />
+      <HeaderCell className={cx(['mock-flowsheet-resultcolumn', 'half-column'])} />
+      <HeaderCell className={cx(['mock-flowsheet-resultcolumn', 'triple-column'])} />
+    </Header>
+    <Body>
+      <Row>
+        <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
+        <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
+        <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
+      </Row>
+      <Row>
+        <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
+        <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
+        <FlowsheetResultCell resultDataSet={multipleResultValues} hideUnit />
+      </Row>
+    </Body>
+  </Table>
+);