+++ packages/terra-clinical-result/src/terra-dev-site/test/clinical-result/result-name-header-cell/SemanticTableWithResultNameHeaderCell.test.jsx
@@ -29,7 +29,8 @@ const defaultBloodPressureResult = [
 ];
 
 const SemanticTableWithResultNameHeaderCell = () => (
-  <Table>
+  // eslint-disable-next-line react/forbid-component-props
+  <Table style={{ tableLayout: 'fixed' }}>
     <Header>
       <ResultNameHeaderCell key="temp" resultName="Temp" unit="degC" />
       <ResultNameHeaderCell key="bp" resultName="BP" unit="mmHg" />