+++ packages/terra-clinical-result/CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Unreleased
 
+* Added
+  * Added back `width: 100%` to flowsheet result cell styles removed in 1.17.0.
+
+* Fixed
+  * Fixed infinite re-render bug.
+
 ## 1.21.0 - (August 2, 2023)
 
-* Changed  
+* Changed
   * Locked `uuid` dependency to `3.4.0` for maximum consistency across Terra packages.
   * Updated translations.